# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop",
        "description": "Pet shop API documentation",
        "contact": {
            "name": "Rabiul Islam",
            "email": "rir.cse.71@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/product/{uuid}": {
            "get": {
                "description": "Get single product",
                "operationId": "6eb40cc9091f67781fe34704490c5c0b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully"
                    },
                    "400": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "description": "Update product",
                "operationId": "9ee612579b1aeb5b1827cbeeb96bfd87",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_uuid",
                                    "price",
                                    "metadata",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "New product"
                                    },
                                    "category_uuid": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "100"
                                    },
                                    "metadata[file_uuid]": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "metadata[brand_uuid]": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Lorem ipsum.."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "400": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "description": "Delete product",
                "operationId": "15d7fa8465a9f928f00d7e145da748b2",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "400": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
        "/file/upload": {
            "post": {
                "description": "Upload a file",
                "operationId": "bfa27d6cac0be37bd45b2cf5008f7995",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "example.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/file/{uuid}": {
            "get": {
                "description": "Get a file",
                "operationId": "90ddfccdb218641f877a64e4419a0efe",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of file",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File retrieved successfully"
                    },
                    "400": {
                        "description": "File not found"
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "description": "Get auth user",
                "operationId": "206bee39a5edcfbaab983f6277f53438",
                "responses": {
                    "200": {
                        "description": "User info retrieved successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/user/edit": {
            "put": {
                "description": "Update user profile",
                "operationId": "223c1619043d5cd966fadc7eca76f646",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "avatar",
                                    "is_marketing",
                                    "address"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "admin@email.com"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123456"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123456"
                                    },
                                    "is_marketing": {
                                        "type": "boolean",
                                        "format": "string",
                                        "example": "0"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Bangladesh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        }
    }
}